Use funcref in quickapp main

+++ src/arcana-incubator/device-abstraction/freefly_camera.cc
@@ -87,7 +87,7 @@ void inc::da::smooth_fps_cam::setup_default_inputs(inc::da::input_manager& input
     input.bindMouseAxis(ge_input_camlook_y, 1, -.65f);
 }
 
-void inc::da::smooth_fps_cam::update_default_inputs(SDL_Window* window, inc::da::input_manager& input, float dt)
+void inc::da::smooth_fps_cam::update_default_inputs(SDL_Window* window, inc::da::input_manager& input, double dt)
 {
     auto speed_mul = 10.f;
