+++ src/arcana-incubator/device-abstraction/freefly_camera.hh
@@ -38,7 +38,7 @@ struct smooth_fps_cam
     // default input all-in-one
 
     void setup_default_inputs(input_manager& input);
-    void update_default_inputs(SDL_Window* window, input_manager& input, float dt);
+    void update_default_inputs(SDL_Window* window, input_manager& input, double dt);
 
 private:
     int _mouse_x_precap = 0;
